+++ Language/Haskell/TH/Desugar/Expand.hs
@@ -29,10 +29,7 @@ module Language.Haskell.TH.Desugar.Expand (
   expandUnsoundly,
 
   -- * Capture-avoiding substitution
-  substTy,
-
-  -- * Free variable calculation
-  freeVarsOfTy
+  substTy
   ) where
 
 import qualified Data.Map as M
@@ -44,10 +41,8 @@ import Control.Applicative
 import Language.Haskell.TH hiding (cxt)
 import Language.Haskell.TH.Syntax ( Quasi(..) )
 import Data.Data
-import qualified Data.Foldable as F
 import Data.Generics
 import Data.List
-import Data.Monoid
 import qualified Data.Traversable as T
 
 import Language.Haskell.TH.Desugar.Core
@@ -306,25 +301,3 @@ expandUnsoundly = expand_ YesIgnore
 -- | Generalization of 'expand' that either can or won't ignore kind annotations.sx
 expand_ :: (DsMonad q, Data a) => IgnoreKinds -> a -> q a
 expand_ ign = everywhereM (mkM (expand_type ign) >=> mkM (expand_pred ign))
-
--- | Compute the free variable of a 'DType'.
-freeVarsOfTy :: DType -> S.Set Name
-freeVarsOfTy = go_ty
-  where
-    go_ty :: DType -> S.Set Name
-    go_ty (DForallT tvbs cxt ty) = (F.foldMap go_tvb tvbs <> go_ty ty <> F.foldMap go_pred cxt)
-                                   S.\\ S.fromList (map extractDTvbName tvbs)
-    go_ty (DAppT t1 t2)          = go_ty t1 <> go_ty t2
-    go_ty (DSigT ty ki)          = go_ty ty <> go_ty ki
-    go_ty (DVarT n)              = S.singleton n
-    go_ty _                      = S.empty
-
-    go_pred :: DPred -> S.Set Name
-    go_pred (DAppPr pr ty) = go_pred pr <> go_ty ty
-    go_pred (DSigPr pr ki) = go_pred pr <> go_ty ki
-    go_pred (DVarPr n)     = S.singleton n
-    go_pred _              = S.empty
-
-    go_tvb :: DTyVarBndr -> S.Set Name
-    go_tvb (DPlainTV{})    = S.empty
-    go_tvb (DKindedTV _ k) = go_ty k